requirements.txt
Django
Markdown
PyYAML
amqp
anyjson
billiard
boto
celery
decorator
django-celery
django-filter
django-nose
djangorestframework
httplib2
ipython
kombu
librabbitmq
nose
pep8
pyparsing
pyrabbit
pysqlite
pytz
readline
redis
requests
wsgiref
django-static-precompiler
flake8
